Rubocopの導入

+++ app/controllers/posts_controller.rb
@@ -51,10 +51,8 @@ class PostsController < ApplicationController
   end
 
   def chat
-    post_user = PostUser.new(post_id: params[:id],user_id: current_user.id)
-    if PostUser.where(post_id: params[:id],user_id: current_user.id) == []
-      post_user.save
-    end
+    post_user = PostUser.new(post_id: params[:id], user_id: current_user.id)
+    post_user.save if PostUser.where(post_id: params[:id], user_id: current_user.id) == []
     @post = Post.find(params[:id])
 
     @message = Message.new
@@ -62,7 +60,7 @@ class PostsController < ApplicationController
     @users = PostUser.where(post_id: @post)
     # ユーザーがbad評価したユーザーがチャットに入っているか
     post_users = PostUser.where(post_id: @post)
-    bad_user_ids = [] 
+    bad_user_ids = []
     post_users.each do |post_user|
       if bad_user = BadUser.find_by(bad_user_id: post_user.user_id, judge_user_id: current_user.id)
         bad_user_ids << bad_user
@@ -76,10 +74,8 @@ class PostsController < ApplicationController
   end
 
   def chat_destroy
-    post_user = PostUser.find_by(post_id: params[:id],user_id: current_user.id)
-      if post_user.present?
-        post_user.destroy
-      end
+    post_user = PostUser.find_by(post_id: params[:id], user_id: current_user.id)
+    post_user.destroy if post_user.present?
     redirect_to root_path
   end
 
@@ -101,5 +97,4 @@ class PostsController < ApplicationController
   def update_params
     params.require(:post).permit(:address, :image, :title, :category_id, :description, :day_time, :prefectures_id, :place, :budget, :name).merge(user_id: current_user.id, id: params[:id])
   end
-
 end
